requirements.txt
orjson>=3.9
psutil>=5.9
pyyaml>=6.0
redis>=5.0
prometheus_client>=0.19
pytest>=7.4
pytest-asyncio>=0.23
pytest-timeout>=2.2
